Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0152-maximum-product-subarray/0152-maximum-product-subarray.java
@@ -7,8 +7,8 @@ class Solution {
         for (int i = 0; i < n; i++) {
             if (pre == 0) pre = 1;
             if (suff == 0) suff = 1;
-            pre *= (double)nums[i];
-            suff *= (double)nums[n - i - 1];
+            pre *= nums[i];
+            suff *= nums[n - i - 1];
             max = Math.max(max, Math.max(pre, suff));
         }
         return max;
